backend/ai-service/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# AI and Computer Vision
torch==2.1.1
torchvision==0.16.1
transformers==4.36.2
diffusers==0.24.0
accelerate==0.25.0
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.4
scipy==1.11.4

# Specific AI models
depth-anything-v2==1.0.0
segment-anything==1.0.0
controlnet-aux==0.0.6

# Image processing
albumentations==1.3.1
imagehash==4.3.1

# Database and caching
asyncpg==0.29.0
redis==5.0.1
qdrant-client==1.7.0

# HTTP and API clients
httpx==0.25.2
aiofiles==23.2.1
requests==2.31.0

# Data validation and serialization
pydantic==2.5.2
pydantic-settings==2.1.0

# Environment and configuration
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Production server
gunicorn==21.2.0

# CUDA support (optional, for GPU acceleration)
# Uncomment if using NVIDIA GPUs
# torch==2.1.1+cu121 -f https://download.pytorch.org/whl/cu121
# torchvision==0.16.1+cu121 -f https://download.pytorch.org/whl/cu121 